Changes the checkbox field.
Makes its type bool.
Adds the `checked` HTML attribute.

+++ cormorant/admin/settings/sanitize/notify_admin_on_confirmation.php
@@ -2,5 +2,5 @@
 
 function notify_admin_on_confirmation($value)
 {
-    return (int) (bool) $value;
+    return (bool) $value;
 }
